+++ core/server/api_container/server/startosis_engine/builtins/time_now_builtin/time_now_builtin.go
@@ -1,4 +1,4 @@
-package time_module_with_now_disabled
+package time_now_builtin
 
 import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
@@ -12,7 +12,6 @@ const (
 )
 
 // GenerateTimeNowBuiltin This only exists to throw a nice interpretation error when print without plan is used
-// This lives here instead of buiitins as we don't add it to kurtosis_module.go
 func GenerateTimeNowBuiltin() func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 	return func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 		return nil, startosis_errors.NewInterpretationError(UseRunPythonInsteadOfTimeNowError)
