+++ core/server/api_container/server/startosis_engine/kurtosis_builtins.go
@@ -2,6 +2,7 @@ package startosis_engine
 
 import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/service_network"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/builtins"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/builtins/import_module"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/builtins/print_builtin"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/builtins/read_file"
@@ -27,8 +28,8 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/runtime_value_store"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_packages"
-	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/time_module_with_now_disabled"
 	starlarkjson "go.starlark.net/lib/json"
+	"go.starlark.net/lib/time"
 	"go.starlark.net/starlark"
 	"go.starlark.net/starlarkstruct"
 )
@@ -40,7 +41,7 @@ func Predeclared() starlark.StringDict {
 		starlarkstruct.Default.GoString(): starlark.NewBuiltin(starlarkstruct.Default.GoString(), starlarkstruct.Make), // extension to build struct in starlark
 
 		// go-starlark time module with time.now() disabled
-		time_module_with_now_disabled.TimeNowBuiltinName: time_module_with_now_disabled.GetTimeModuleWithNowDisabled(),
+		time.Module.Name: builtins.TimeModuleWithNowDisabled(),
 	}
 }
