+++ core/server/api_container/server/startosis_engine/startosis_interpreter.go
@@ -345,7 +345,6 @@ func (interpreter *StartosisInterpreter) buildBindings(thread *starlark.Thread, 
 	}
 
 	predeclared := Predeclared()
-
 	// Add custom Kurtosis module
 	predeclared[builtins.KurtosisModuleName] = kurtosisModule
